requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.3.4
cycler==0.12.1
fastapi==0.129.0
filelock==3.24.3
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
greenlet==3.3.1
h11==0.16.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mediapipe==0.10.32
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.1
polars==1.38.1
polars-runtime-32==1.38.1
psutil==7.2.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.5
scipy==1.17.0
setuptools==82.0.0
six==1.17.0
sounddevice==0.5.5
SQLAlchemy==2.0.46
starlette==0.52.1
sympy==1.14.0
torch==2.10.0
torchvision==0.25.0
triton==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.14
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.41.0
